refactor: move request_builder under http for organization
