+++ pinterest_dl/low_level/api/pinterest_api.py
@@ -11,7 +11,7 @@ from pinterest_dl.exceptions import (
 )
 from pinterest_dl.low_level.api.endpoints import Endpoint
 from pinterest_dl.low_level.api.pinterest_response import PinResponse
-from pinterest_dl.low_level.ops.request_builder import RequestBuilder
+from pinterest_dl.low_level.http.request_builder import RequestBuilder
 
 
 class PinterestAPI:
